email/requirements.txt
annotated-types==0.7.0
blessed==1.20.0
botocore==1.35.51
bump2version==1.0.1
bumpversion==0.6.0
cffi==1.17.1
chalice==1.31.2
click==8.1.7
cryptography==43.0.3
dnspython==2.7.0
inquirer==2.10.1
Jinja2==3.1.4
jmespath==1.0.1
MarkupSafe==3.0.2
pycparser==2.22
pydantic==2.6.4
pydantic_core==2.16.3
pymongo==4.6.3
pymongoose==1.3.8
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-editor==1.0.4
PyYAML==6.0.2
readchar==4.2.0
secure-smtplib==0.1.1
six==1.16.0
typing_extensions==4.12.2
urllib3==2.2.3
wcwidth==0.2.13
